requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
openai==1.51.0
pillow==10.4.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
pytest==8.3.3
httpx==0.27.2
